Created Map Page
Currently the map is a canvas element that loads in that map upon loading. THe idea is to add dynamically placed dots that are created based on the data inside the database. Will do this using a 1to1 mapping of lets say Arecibo to x,y coords in the canvas. This way any cave added to the database outside of the already existing ones will be able to automatically place unto the map.